+++ src/Resources/contao/dca/tl_content.php
@@ -1,26 +1,26 @@
 <?php
 
-$arrDca = &$GLOBALS['TL_DCA']['tl_content'];
+$dca = &$GLOBALS['TL_DCA']['tl_content'];
 
 /**
  * Callbacks
  */
-$arrDca['config']['onsubmit_callback']['addNewsPaginationStopElement'] =
-    ['HeimrichHannot\NewsPagination\Backend\Content', 'addNewsPaginationStopElement'];
+$dca['config']['onsubmit_callback']['addNewsPaginationStopElement'] =
+    [\HeimrichHannot\NewsPaginationBundle\DataContainer\ContentContainer::class, 'addNewsPaginationStopElement'];
 
 /**
  * Palettes
  */
-$arrDca['palettes'][\HeimrichHannot\NewsPaginationBundle\NewsPaginationBundle::CONTENT_ELEMENT_NEWS_PAGINATION_START] =
+$dca['palettes'][\HeimrichHannot\NewsPaginationBundle\NewsPaginationBundle::CONTENT_ELEMENT_NEWS_PAGINATION_START] =
     '{type_legend},type,newsPaginationTitle;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID,space;{invisible_legend:hide},invisible,start,stop';
-$arrDca['palettes'][\HeimrichHannot\NewsPaginationBundle\NewsPaginationBundle::CONTENT_ELEMENT_NEWS_PAGINATION_STOP]  =
+$dca['palettes'][\HeimrichHannot\NewsPaginationBundle\NewsPaginationBundle::CONTENT_ELEMENT_NEWS_PAGINATION_STOP]  =
     '{type_legend},type;{template_legend:hide},customTpl;{protected_legend:hide},protected;{invisible_legend:hide},invisible,start,stop';
 
 
 /**
  * Fields
  */
-$arrFields = [
+$fields = [
     'newsPaginationTitle' => [
         'label'                   => &$GLOBALS['TL_LANG']['tl_content']['newsPaginationTitle'],
         'exclude'                 => true,
@@ -35,4 +35,4 @@ $arrFields = [
     ]
 ];
 
-$arrDca['fields'] += $arrFields;
+$dca['fields'] = array_merge(is_array($dca['fields']) ? $dca['fields'] : [], $fields);