+++ src/Resources/contao/languages/de/tl_reader_config.php
@@ -1,0 +1,10 @@
+<?php
+
+if (class_exists('HeimrichHannot\ReaderBundle\HeimrichHannotContaoReaderBundle')) {
+    $lang = &$GLOBALS['TL_LANG']['tl_reader_config'];
+
+    /**
+     * Legends
+     */
+    $lang['pagination_legend'] = 'Paginierung';
+}